Create README.md
Our project is a webpage showcasing the Netflix platform and its capabilities , created using HTML , CSS and JavaScript. The page has a logo in the header , a sign-in button , a language picker, and a button to begin a membership. The primary part is made up of numerous containers that contain text, graphics, and videos that describe the functions of Netflix, including watching on TV, downloading content to watch later, streaming on various devices and setting up kid-specific accounts. The user can extend the response by selecting a button on the page’s part with commonly asked questions, which is also included. Bootstrap is used on the page for responsive design and style.